Rename opt_models and register as app
