+++ offgridplanner/optimization/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ProjectsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "offgridplanner.optimization"
