+++ offgridplanner/optimization/base_optimizer.py
