+++ offgridplanner/optimization/solar_potential.py
